+++ Samples~/DefinitionWorkflow/Definitions/Editor/Definition_MediaSampleTone.definition.cs
@@ -1,0 +1,22 @@
+// <deucarian-definition schema="media" />
+// Editable declaration. Use the package definition editor or edit the values below.
+namespace Deucarian.ProjectDefinitions.Definition_media
+{
+    public static class Definition_MediaSampleTone
+    {
+        public static global::Deucarian.Media.Editor.Definitions.MediaDefinitionSpec Value =>
+        // definition-value
+        new global::Deucarian.Media.Editor.Definitions.MediaDefinitionSpec
+        {
+            AudioType = global::UnityEngine.AudioType.UNKNOWN,
+            Clip = global::Deucarian.Editor.Definitions.DeucarianDefinitionAssets.Load<global::UnityEngine.AudioClip>("9f85953892a55f844852e797672aa41a", 8300000L),
+            Id = "311442c302a94fbeab2dfb64f3f27f56",
+            Kind = global::Deucarian.Media.MediaKind.Audio,
+            Loop = true,
+            Name = "MediaSampleTone",
+            Url = "",
+            Volume = 0.3f,
+        };
+        // end-definition-value
+    }
+}